evalbot/pyproject.toml
[project]
name = "evalbot"
version = "1.0.0"
description = "Evalbot MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp==1.9.4",
    "PyYAML>=6.0.3",
    "requests>=2.32.5"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
mcp-server-evalbot = "evalbot.server:main"
